refactor: DRY types, add Zod validation, fix SKIP_ENV_VALIDATION gaps
- Separate pure type interfaces into dedicated types.ts files
- Add Zod schemas for corpus and infra config validation at parse boundaries
- Derive SeedObservation/SeedCorpus types from Zod schemas via z.infer
- Extract duplicated Pinecone env setup into shared env-setup.ts
- Add SKIP_ENV_VALIDATION to @vendor/embed and @vendor/upstash env.ts
- Remove declared-but-unimplemented ./scorers export from package.json
